Moved ToastProvider to services and Toast component to components
