+++ common/v2/services/index.ts
@@ -1,5 +1,6 @@
 export * from './ApiService';
 export * from './EthService';
 export * from './Store';
+export * from './Toasts';
 export { RatesContext, RatesProvider } from './RatesProvider';
 export { DevToolsContext, DevToolsProvider, useDevTools } from './DevToolsProvider';
